ENG-7: Move table components into table folder
