+++ client/components/Table/PlayersTable.tsx
